create load package
