+++ src/internal/packager2/load/import_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2021-Present The Zarf Authors
 
-package create
+package load
 
 import (
 	"os"
