+++ src/internal/packager2/load/testdata/import/flavor/expected.yaml
