+++ src/internal/packager2/load/testdata/import/import-each-other/expected.yaml
