+++ src/internal/packager2/load/testdata/import/variables/expected.yaml
